+++ tests/assets/dashAgGridFunctions.js
@@ -172,3 +172,186 @@ function monthToComparableNumber(date) {
   const dayNumber = parseInt(date.split('/')[0]);
   return yearNumber * 10000 + monthNumber * 100 + dayNumber;
 }
+
+const {useImperativeHandle, useState, useEffect, forwardRef} = React;
+
+// This example was adapted from https://www.ag-grid.com/react-data-grid/component-filter/
+// The only differences are:
+// - React.createElement instead of JSX
+// - setProps, which all Dash components use to report user interactions,
+//     instead of a plain js event handler
+dagfuncs.YearFilter = forwardRef((props, ref) => {
+   const [year, setYear] = useState('All');
+
+   useImperativeHandle(ref, () => {
+       return {
+           doesFilterPass(params) {
+               return params.data.year >= 2010;
+           },
+
+           isFilterActive() {
+               return year === '2010'
+           },
+
+           // this example isn't using getModel() and setModel(),
+           // so safe to just leave these empty. don't do this in your code!!!
+           getModel() {
+           },
+
+           setModel() {
+           }
+       }
+   });
+
+   useEffect(() => {
+       props.filterChangedCallback()
+   }, [year]);
+
+    setProps = ({value}) => {
+        if (value) {
+            setYear(value)
+        }
+    }
+
+    return React.createElement(
+        window.dash_core_components.RadioItems,
+        {
+            options:[
+                {'label': 'All', 'value': 'All'},
+                {'label': 'Since 2010', 'value': '2010'},
+            ],
+            value: year,
+            setProps
+        }
+        )
+});
+
+dagfuncs.setBody = () => {
+    return document.querySelector('body')
+}
+
+// cell editor custom component  - dmc.Select
+dagfuncs.DMC_Select = class {
+    // gets called once before the renderer is used
+    init(params) {
+        // create the cell
+        this.params = params;
+
+        // function for when Dash is trying to send props back to the component / server
+        var setProps = (props) => {
+            if (typeof props.value != typeof undefined) {
+                // updates the value of the editor
+                this.value = props.value;
+
+                // re-enables keyboard event
+                delete params.colDef.suppressKeyboardEvent;
+
+                // tells the grid to stop editing the cell
+                params.api.stopEditing();
+
+                // sets focus back to the grid's previously active cell
+                this.prevFocus.focus();
+            }
+        };
+        this.eInput = document.createElement('div');
+
+        // renders component into the editor element
+        ReactDOM.render(
+            React.createElement(window.dash_mantine_components.Select, {
+                data: params.options,
+                value: params.value,
+                setProps,
+                style: {width: params.column.actualWidth-2,  ...params.style},
+                className: params.className,
+                clearable: params.clearable,
+                searchable: params.searchable || true,
+                creatable: params.creatable,
+                debounce: params.debounce,
+                disabled: params.disabled,
+                filterDataOnExactSearchMatch:
+                    params.filterDataOnExactSearchMatch,
+                limit: params.limit,
+                maxDropdownHeight: params.maxDropdownHeight,
+                nothingFound: params.nothingFound,
+                placeholder: params.placeholder,
+                required: params.required,
+                searchValue: params.searchValue,
+                shadow: params.shadow,
+                size: params.size,
+                styles: params.styles,
+                switchDirectionOnFlip: params.switchDirectionOnFlip,
+                variant: params.variant,
+            }),
+            this.eInput
+        );
+
+        // allows focus event
+        this.eInput.tabIndex = '0';
+
+        // sets editor value to the value from the cell
+        this.value = params.value;
+    }
+
+    // gets called once when grid ready to insert the element
+    getGui() {
+        return this.eInput;
+    }
+
+    focusChild() {
+        // needed to delay and allow the component to render
+        setTimeout(() => {
+            var inp = this.eInput.getElementsByClassName(
+                'mantine-Select-input'
+            )[0];
+            inp.tabIndex = '1';
+
+            // disables keyboard event
+            this.params.colDef.suppressKeyboardEvent = (params) => {
+                const gridShouldDoNothing = params.editing;
+                return gridShouldDoNothing;
+            };
+            // shows dropdown options
+            inp.focus();
+        }, 100);
+    }
+
+    // focus and select can be done after the gui is attached
+    afterGuiAttached() {
+        // stores the active cell
+        this.prevFocus = document.activeElement;
+
+        // adds event listener to trigger event to go into dash component
+        this.eInput.addEventListener('focus', this.focusChild());
+
+        // triggers focus event
+        this.eInput.focus();
+    }
+
+    // returns the new value after editing
+    getValue() {
+        return this.value;
+    }
+
+    // any cleanup we need to be done here
+    destroy() {
+        // sets focus back to the grid's previously active cell
+        this.prevFocus.focus();
+    }
+};
+
+dagfuncs.contextTest = (params) => {
+    var result = [
+      {
+        // custom item
+        name: 'Alert ' + params.value,
+        action: () => {
+          window.alert('Alerting about ' + params.value);
+        },
+        cssClasses: ['redFont', 'bold'],
+      },
+      'copy',
+      'separator',
+      'chartRange',
+    ];
+    return result;
+};